requirements.txt
# Project dependencies 
python-telegram-bot
python-dotenv
google-generativeai